pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "kmtronic-relay"
version = "1.0.0"
description = "Python API for KMTronic USB relay boards"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Marvin Sass", email = "marvin.sass@dismail.de"}
]
keywords = ["kmtronic", "relay", "usb", "hardware", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: System :: Hardware :: Hardware Drivers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pyserial>=3.5",
]

[project.urls]
Homepage = "https://github.com/HWS-XMS/KMtronicUSBRelay"
"Bug Reports" = "https://github.com/HWS-XMS/KMtronicUSBRelay/issues"
Source = "https://github.com/HWS-XMS/KMtronicUSBRelay"
